+++ libraries/ui-library/src/utils/animation.ts
@@ -40,3 +40,22 @@ export function flipAnimate(
     options
   );
 }
+
+/**
+ * Animates an element using keyframes. Returns a promise that resolves after the animation completes or gets canceled.
+ */
+export function animateTo(el: HTMLElement, keyframes: Keyframe[], options?: KeyframeAnimationOptions) {
+  return new Promise((resolve) => {
+    if (options?.duration === Infinity) {
+      throw new Error('Promise-based animations must be finite.');
+    }
+
+    const animation = el.animate(keyframes, {
+      ...options,
+      duration: options!.duration,
+    });
+
+    animation.addEventListener('cancel', resolve, { once: true });
+    animation.addEventListener('finish', resolve, { once: true });
+  });
+}